fix 'select category shortcut', code cleanup, #114
