+++ src/store/selectors/annotationCategorySelector.ts
@@ -1,0 +1,10 @@
+import { ImageViewer } from "../../types/ImageViewer";
+import { Category } from "../../types/Category";
+
+export const annotationCategorySelector = ({
+  imageViewer,
+}: {
+  imageViewer: ImageViewer;
+}): Array<Category> => {
+  return imageViewer.categories;
+};
